rtsp/requirements.txt
numpy==1.21.6
opencv-contrib-python==4.6.0.66
Pillow==9.2.0
